+++ tests/testthat/test-warehousing-duckdb.R
@@ -220,8 +220,8 @@ test_that("Ramses on DuckDB (system test)", {
       patient_id = "1555756339",
       therapy_id = "592a738e4c2afcae6f625c01856151e0",
       antiinfective_type = "antibacterial",
-      therapy_start = "2016-08-01 11:15:19+01:00",
-      therapy_end = "2016-08-03 11:15:19+01:00"
+      therapy_start = as.POSIXct("2016-08-01 11:15:19", tz = Sys.timezone(), origin = "1960-01-01"),
+      therapy_end = as.POSIXct("2016-08-03 11:15:19", tz = Sys.timezone(), origin = "1960-01-01")
     )
   )
   
@@ -242,10 +242,10 @@ test_that("Ramses on DuckDB (system test)", {
       patient_id = "1555756339",
       therapy_id = "592a738e4c2afcae6f625c01856151e0",
       antiinfective_type = "antibacterial",
-      therapy_start = "2016-08-01 11:15:19+01:00",
-      therapy_end = "2016-08-03 11:15:19+01:00"
+      therapy_start = as.POSIXct("2016-08-01 11:15:19", tz = Sys.timezone(), origin = "1960-01-01"),
+      therapy_end = as.POSIXct("2016-08-03 11:15:19", tz = Sys.timezone(), origin = "1960-01-01")
     )
-    )
+  )
   
   # other database functions --------------------------------------------
   
@@ -308,25 +308,25 @@ test_that("Ramses on DuckDB (system test)", {
     patient_id = "99999999999",
     therapy_id = "5528fc41106bb48eb4d48bc412e13e67",
     therapy_start = structure(c(1438939620, 1438939620, 1438939620, 1438939620, 
-                                1438939620, 1438939620), tzone = "", class = c("POSIXct", "POSIXt")), 
+                                1438939620, 1438939620), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     therapy_end = structure(c(1439810400, 1439810400, 1439810400, 1439810400,
-                              1439810400, 1439810400), tzone = "", class = c("POSIXct", "POSIXt")), 
+                              1439810400, 1439810400), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     t_start = structure(c(1438939620, 1438943220, 1438946820, 1438950420, 
-                          1438954020, 1438957620), tzone = "", class = c("POSIXct", "POSIXt")), 
+                          1438954020, 1438957620), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     t_end = structure(c(1438943220, 1438946820, 1438950420, 1438954020, 
-                        1438957620, 1438961220), tzone = "", class = c("POSIXct", "POSIXt")), 
+                        1438957620, 1438961220), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     parenteral = 1L
   )
   test_expected_tail <- dplyr::tibble(
     t = 236:241,
     patient_id = "99999999999",
     therapy_id = "5528fc41106bb48eb4d48bc412e13e67",
-    therapy_start = structure(1438939620, tzone = "", class = c("POSIXct", "POSIXt")),
-    therapy_end = structure(1439810400, tzone = "", class = c("POSIXct", "POSIXt")),
+    therapy_start = structure(1438939620, tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")),
+    therapy_end = structure(1439810400, tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")),
     t_start = structure(c(1439789220, 1439792820, 1439796420, 1439800020, 
-                          1439803620, 1439807220), tzone = "", class = c("POSIXct", "POSIXt")),
+                          1439803620, 1439807220), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")),
     t_end = structure(c(1439792820, 1439796420, 1439800020, 1439803620, 1439807220, 
-                        1439810400), tzone = "", class = c("POSIXct", "POSIXt")),
+                        1439810400), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")),
     parenteral = 0L
   )
   
@@ -362,16 +362,16 @@ test_that("Ramses on DuckDB (system test)", {
     therapy_id = "f770855cf9d424c76fdfbc9786d508ac", 
     therapy_start = structure(
       c(1444239793, 1444239793, 1444239793, 1444239793, 
-        1444239793, 1444239793), tzone = "", class = c("POSIXct", "POSIXt")), 
+        1444239793, 1444239793), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     therapy_end = structure(
       c(1444681333, 1444681333, 1444681333, 1444681333, 
-        1444681333, 1444681333), tzone = "", class = c("POSIXct", "POSIXt")), 
+        1444681333, 1444681333), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     t_start = structure(
       c(1444660993, 1444664593, 1444668193, 1444671793, 
-        1444675393, 1444678993), tzone = "", class = c("POSIXct", "POSIXt")), 
+        1444675393, 1444678993), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     t_end = structure(
       c(1444664593, 1444668193, 1444671793, 1444675393, 
-        1444678993, 1444681333), tzone = "", class = c("POSIXct", "POSIXt")), 
+        1444678993, 1444681333), tzone = Sys.timezone(), class = c("POSIXct", "POSIXt")), 
     parenteral = 0L
   )
   expect_equal(head(therapy_table(test_episode, collect = TRUE)), 
@@ -681,10 +681,10 @@ test_that("Ramses on DuckDB (system test)", {
     utils::capture.output(TherapyEpisode(db_conn, "89ac870bc1c1e4b2a37cec79d188cb08"))[1:8],
     c("TherapyEpisode 89ac870bc1c1e4b2a37cec79d188cb08 ", "Patient:   1555756339 ", 
       paste0("Start:     ",
-             as.character(as.POSIXct("2017-07-02 01:15:46", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2017-07-02 01:15:46", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "), 
       paste0("End:       ",
-             as.character(as.POSIXct("2017-07-06 01:35:46", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2017-07-06 01:35:46", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "), 
       "", "Medications:", "  > Amoxicillin/clavulanic acid IV 1.2g 2 days", 
       "  > Clarithromycin ORAL 500mg 4 days"))
@@ -692,10 +692,10 @@ test_that("Ramses on DuckDB (system test)", {
     utils::capture.output(TherapyEpisode(db_conn, "fa179f4bcf3efa1e21225ab207ab40c4"))[1:11],
     c("TherapyEpisode fa179f4bcf3efa1e21225ab207ab40c4 ", "Patient:   3422481921 ", 
       paste0("Start:     ", 
-             as.character(as.POSIXct("2017-11-15 15:33:36", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2017-11-15 15:33:36", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "),
       paste0("End:       ",
-             as.character(as.POSIXct("2017-12-01 21:11:36", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2017-12-01 21:11:36", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "),
       "", "Medications:", "  > Amoxicillin/clavulanic acid IV 1.2g 2 days", 
       "  > Amoxicillin/clavulanic acid IV 1.2g 2 days", "  > Piperacillin/tazobactam IV 4.5g 4 days", 
@@ -722,10 +722,10 @@ test_that("Ramses on DuckDB (system test)", {
     c("MedicationRequest 5528fc41106bb48eb4d48bc412e13e67 ", "Clarithromycin IV 500mg 0 days ", 
       "Patient:     99999999999 ", 
       paste0("Start:        ", 
-             as.character(as.POSIXct("2015-08-07 10:27:00", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2015-08-07 10:27:00", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "),
       paste0("End:          ", 
-             as.character(as.POSIXct("2015-08-07 15:59:00", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2015-08-07 15:59:00", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "),
       "Therapy:      5528fc41106bb48eb4d48bc412e13e67 ", 
       "", "Database connection:"))
@@ -734,10 +734,10 @@ test_that("Ramses on DuckDB (system test)", {
     c("MedicationRequest 1ab55e515af6b86dde76abbe0bffbd3f ", "Clarithromycin ORAL 500mg 4 days ", 
       "Patient:     3894468747 ", 
       paste0("Start:        ", 
-             as.character(as.POSIXct("2015-10-01 21:38:55", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2015-10-01 21:38:55", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "), 
       paste0("End:          ", 
-             as.character(as.POSIXct("2015-10-05 21:38:55", tz = "Europe/London"), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
+             as.character(as.POSIXct("2015-10-05 21:38:55", tz = Sys.timezone()), tz = Sys.timezone(), format = "%Y-%m-%d %H:%M:%S %Z"),
              " "),
       "Combination:  1ab55e515af6b86dde76abbe0bffbd3f ", 
       "Therapy:      1ab55e515af6b86dde76abbe0bffbd3f ", "", "Database connection:"